feat: replace the in-memory cache with the new framework
