+++ Cargo.toml
@@ -7,9 +7,7 @@ members = [
     "foyer-cli",
     "foyer-common",
     "foyer-intrusive",
-    "foyer-intrusive-v2",
     "foyer-memory",
-    "foyer-memory-v2",
     "foyer-storage",
     "foyer-util",
 ]
@@ -53,7 +51,6 @@ tokio = { package = "madsim-tokio", version = "0.2", features = [
 # foyer components
 foyer-common = { version = "0.13.0-dev", path = "foyer-common" }
 foyer-intrusive = { version = "0.13.0-dev", path = "foyer-intrusive" }
-foyer-intrusive-v2 = { version = "0.13.0-dev", path = "foyer-intrusive-v2" }
 foyer-memory = { version = "0.13.0-dev", path = "foyer-memory" }
 foyer-storage = { version = "0.13.0-dev", path = "foyer-storage" }
 foyer = { version = "0.13.0-dev", path = "foyer" }
