+++ foyer-memory-bakcup/benches/bench_dynamic_dispatch.rs
