+++ foyer-memory-bakcup/src/cache.rs
@@ -12,14 +12,9 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-use std::{
-    fmt::Debug,
-    future::Future,
-    hash::{Hash, RandomState},
-    ops::Deref,
-    sync::Arc,
-};
+use std::{fmt::Debug, hash::Hash, ops::Deref, sync::Arc};
 
+use ahash::RandomState;
 use equivalent::Equivalent;
 use foyer_common::{
     code::{HashBuilder, Key, Value},
@@ -27,36 +22,57 @@ use foyer_common::{
     future::Diversion,
     runtime::SingletonHandle,
 };
+use futures::Future;
 use pin_project::pin_project;
 use serde::{Deserialize, Serialize};
 use tokio::sync::oneshot;
 
 use crate::{
+    context::CacheContext,
     eviction::{
-        fifo::{Fifo, FifoConfig},
-        lfu::{Lfu, LfuConfig},
-        lru::{Lru, LruConfig},
-        s3fifo::{S3Fifo, S3FifoConfig},
+        fifo::{Fifo, FifoHandle},
+        lfu::{Lfu, LfuHandle},
+        lru::{Lru, LruHandle},
+        s3fifo::{S3Fifo, S3FifoHandle},
+        sanity::SanityEviction,
     },
-    raw::{FetchMark, FetchState, RawCache, RawCacheConfig, RawCacheEntry, RawFetch, Weighter},
-    record::CacheHint,
+    generic::{FetchMark, FetchState, GenericCache, GenericCacheConfig, GenericCacheEntry, GenericFetch, Weighter},
+    indexer::{hash_table::HashTableIndexer, sanity::SanityIndexer},
+    FifoConfig, LfuConfig, LruConfig, S3FifoConfig,
 };
 
-pub type FifoCache<K, V, S = RandomState> = RawCache<Fifo<K, V>, S>;
-pub type FifoCacheEntry<K, V, S = RandomState> = RawCacheEntry<Fifo<K, V>, S>;
-pub type FifoFetch<K, V, ER, S = RandomState> = RawFetch<Fifo<K, V>, ER, S>;
+pub type FifoCache<K, V, S = RandomState> =
+    GenericCache<K, V, SanityEviction<Fifo<(K, V)>>, SanityIndexer<HashTableIndexer<K, FifoHandle<(K, V)>>>, S>;
+pub type FifoCacheEntry<K, V, S = RandomState> =
+    GenericCacheEntry<K, V, SanityEviction<Fifo<(K, V)>>, SanityIndexer<HashTableIndexer<K, FifoHandle<(K, V)>>>, S>;
+pub type FifoFetch<K, V, ER, S = RandomState> =
+    GenericFetch<K, V, SanityEviction<Fifo<(K, V)>>, SanityIndexer<HashTableIndexer<K, FifoHandle<(K, V)>>>, S, ER>;
 
-pub type S3FifoCache<K, V, S = RandomState> = RawCache<S3Fifo<K, V>, S>;
-pub type S3FifoCacheEntry<K, V, S = RandomState> = RawCacheEntry<S3Fifo<K, V>, S>;
-pub type S3FifoFetch<K, V, ER, S = RandomState> = RawFetch<S3Fifo<K, V>, ER, S>;
+pub type LruCache<K, V, S = RandomState> =
+    GenericCache<K, V, SanityEviction<Lru<(K, V)>>, SanityIndexer<HashTableIndexer<K, LruHandle<(K, V)>>>, S>;
+pub type LruCacheEntry<K, V, S = RandomState> =
+    GenericCacheEntry<K, V, SanityEviction<Lru<(K, V)>>, SanityIndexer<HashTableIndexer<K, LruHandle<(K, V)>>>, S>;
+pub type LruFetch<K, V, ER, S = RandomState> =
+    GenericFetch<K, V, SanityEviction<Lru<(K, V)>>, SanityIndexer<HashTableIndexer<K, LruHandle<(K, V)>>>, S, ER>;
 
-pub type LruCache<K, V, S = RandomState> = RawCache<Lru<K, V>, S>;
-pub type LruCacheEntry<K, V, S = RandomState> = RawCacheEntry<Lru<K, V>, S>;
-pub type LruFetch<K, V, ER, S = RandomState> = RawFetch<Lru<K, V>, ER, S>;
+pub type LfuCache<K, V, S = RandomState> =
+    GenericCache<K, V, SanityEviction<Lfu<(K, V)>>, SanityIndexer<HashTableIndexer<K, LfuHandle<(K, V)>>>, S>;
+pub type LfuCacheEntry<K, V, S = RandomState> =
+    GenericCacheEntry<K, V, SanityEviction<Lfu<(K, V)>>, SanityIndexer<HashTableIndexer<K, LfuHandle<(K, V)>>>, S>;
+pub type LfuFetch<K, V, ER, S = RandomState> =
+    GenericFetch<K, V, SanityEviction<Lfu<(K, V)>>, SanityIndexer<HashTableIndexer<K, LfuHandle<(K, V)>>>, S, ER>;
 
-pub type LfuCache<K, V, S = RandomState> = RawCache<Lfu<K, V>, S>;
-pub type LfuCacheEntry<K, V, S = RandomState> = RawCacheEntry<Lfu<K, V>, S>;
-pub type LfuFetch<K, V, ER, S = RandomState> = RawFetch<Lfu<K, V>, ER, S>;
+pub type S3FifoCache<K, V, S = RandomState> =
+    GenericCache<K, V, SanityEviction<S3Fifo<(K, V)>>, SanityIndexer<HashTableIndexer<K, S3FifoHandle<(K, V)>>>, S>;
+pub type S3FifoCacheEntry<K, V, S = RandomState> = GenericCacheEntry<
+    K,
+    V,
+    SanityEviction<S3Fifo<(K, V)>>,
+    SanityIndexer<HashTableIndexer<K, S3FifoHandle<(K, V)>>>,
+    S,
+>;
+pub type S3FifoFetch<K, V, ER, S = RandomState> =
+    GenericFetch<K, V, SanityEviction<S3Fifo<(K, V)>>, SanityIndexer<HashTableIndexer<K, S3FifoHandle<(K, V)>>>, S, ER>;
 
 /// A cached entry holder of the in-memory cache.
 #[derive(Debug)]
@@ -68,12 +84,12 @@ where
 {
     /// A cached entry holder of the in-memory FIFO cache.
     Fifo(FifoCacheEntry<K, V, S>),
-    /// A cached entry holder of the in-memory S3FIFO cache.
-    S3Fifo(S3FifoCacheEntry<K, V, S>),
     /// A cached entry holder of the in-memory LRU cache.
     Lru(LruCacheEntry<K, V, S>),
     /// A cached entry holder of the in-memory LFU cache.
     Lfu(LfuCacheEntry<K, V, S>),
+    /// A cached entry holder of the in-memory S3FIFO cache.
+    S3Fifo(S3FifoCacheEntry<K, V, S>),
 }
 
 impl<K, V, S> Clone for CacheEntry<K, V, S>
@@ -190,13 +206,13 @@ where
         }
     }
 
-    /// Hint of the cached entry.
-    pub fn hint(&self) -> CacheHint {
+    /// Context of the cached entry.
+    pub fn context(&self) -> CacheContext {
         match self {
-            CacheEntry::Fifo(entry) => entry.hint().clone().into(),
-            CacheEntry::Lru(entry) => entry.hint().clone().into(),
-            CacheEntry::Lfu(entry) => entry.hint().clone().into(),
-            CacheEntry::S3Fifo(entry) => entry.hint().clone().into(),
+            CacheEntry::Fifo(entry) => entry.context().clone().into(),
+            CacheEntry::Lru(entry) => entry.context().clone().into(),
+            CacheEntry::Lfu(entry) => entry.context().clone().into(),
+            CacheEntry::S3Fifo(entry) => entry.context().clone().into(),
         }
     }
 
@@ -211,7 +227,7 @@ where
     }
 
     /// External reference count of the cached entry.
-    pub fn refs(&self) -> isize {
+    pub fn refs(&self) -> usize {
         match self {
             CacheEntry::Fifo(entry) => entry.refs(),
             CacheEntry::Lru(entry) => entry.refs(),
@@ -236,23 +252,17 @@ where
 pub enum EvictionConfig {
     /// FIFO eviction algorithm config.
     Fifo(FifoConfig),
-    /// S3FIFO eviction algorithm config.
-    S3Fifo(S3FifoConfig),
     /// LRU eviction algorithm config.
     Lru(LruConfig),
     /// LFU eviction algorithm config.
     Lfu(LfuConfig),
+    /// S3FIFO eviction algorithm config.
+    S3Fifo(S3FifoConfig),
 }
 
 impl From<FifoConfig> for EvictionConfig {
     fn from(value: FifoConfig) -> EvictionConfig {
         EvictionConfig::Fifo(value)
-    }
-}
-
-impl From<S3FifoConfig> for EvictionConfig {
-    fn from(value: S3FifoConfig) -> EvictionConfig {
-        EvictionConfig::S3Fifo(value)
     }
 }
 
@@ -268,6 +278,12 @@ impl From<LfuConfig> for EvictionConfig {
     }
 }
 
+impl From<S3FifoConfig> for EvictionConfig {
+    fn from(value: S3FifoConfig) -> EvictionConfig {
+        EvictionConfig::S3Fifo(value)
+    }
+}
+
 /// In-memory cache builder.
 pub struct CacheBuilder<K, V, S>
 where
@@ -280,9 +296,7 @@ where
     capacity: usize,
     shards: usize,
     eviction_config: EvictionConfig,
-
-    slab_initial_capacity: usize,
-    slab_segment_size: usize,
+    object_pool_capacity: usize,
 
     hash_builder: S,
     weighter: Arc<dyn Weighter<K, V>>,
@@ -302,11 +316,14 @@ where
 
             capacity,
             shards: 8,
-            eviction_config: LruConfig::default().into(),
-
-            slab_initial_capacity: 64 * 1024,
-            slab_segment_size: 16 * 1024,
-
+            eviction_config: LfuConfig {
+                window_capacity_ratio: 0.1,
+                protected_capacity_ratio: 0.8,
+                cmsketch_eps: 0.001,
+                cmsketch_confidence: 0.9,
+            }
+            .into(),
+            object_pool_capacity: 1024,
             hash_builder: RandomState::default(),
             weighter: Arc::new(|_, _| 1),
             event_listener: None,
@@ -345,19 +362,13 @@ where
         self
     }
 
-    /// Set slab initial capacity for each shard.
+    /// Set object pool for handles. The object pool is used to reduce handle allocation.
     ///
-    /// The default value is 64 KiB.
-    pub fn with_slab_initial_capacity(mut self, slab_initial_capacity: usize) -> Self {
-        self.slab_initial_capacity = slab_initial_capacity;
-        self
-    }
-
-    /// Set slab segment size for each shard.
+    /// The optimized value is supposed to be equal to the max cache entry count.
     ///
-    /// The default value is 16 KiB.
-    pub fn with_slab_segment_size(mut self, slab_segment_size: usize) -> Self {
-        self.slab_segment_size = slab_segment_size;
+    /// The default value is 1024.
+    pub fn with_object_pool_capacity(mut self, object_pool_capacity: usize) -> Self {
+        self.object_pool_capacity = object_pool_capacity;
         self
     }
 
@@ -371,8 +382,7 @@ where
             capacity: self.capacity,
             shards: self.shards,
             eviction_config: self.eviction_config,
-            slab_initial_capacity: self.slab_initial_capacity,
-            slab_segment_size: self.slab_segment_size,
+            object_pool_capacity: self.object_pool_capacity,
             hash_builder,
             weighter: self.weighter,
             event_listener: self.event_listener,
@@ -402,46 +412,42 @@ where
         }
 
         match self.eviction_config {
-            EvictionConfig::Fifo(eviction_config) => Cache::Fifo(Arc::new(RawCache::new(RawCacheConfig {
+            EvictionConfig::Fifo(eviction_config) => Cache::Fifo(Arc::new(GenericCache::new(GenericCacheConfig {
                 name: self.name,
                 capacity: self.capacity,
                 shards: self.shards,
                 eviction_config,
-                slab_initial_capacity: self.slab_initial_capacity,
-                slab_segment_size: self.slab_segment_size,
+                object_pool_capacity: self.object_pool_capacity,
                 hash_builder: self.hash_builder,
                 weighter: self.weighter,
                 event_listener: self.event_listener,
             }))),
-            EvictionConfig::S3Fifo(eviction_config) => Cache::S3Fifo(Arc::new(RawCache::new(RawCacheConfig {
+            EvictionConfig::Lru(eviction_config) => Cache::Lru(Arc::new(GenericCache::new(GenericCacheConfig {
                 name: self.name,
                 capacity: self.capacity,
                 shards: self.shards,
                 eviction_config,
-                slab_initial_capacity: self.slab_initial_capacity,
-                slab_segment_size: self.slab_segment_size,
+                object_pool_capacity: self.object_pool_capacity,
                 hash_builder: self.hash_builder,
                 weighter: self.weighter,
                 event_listener: self.event_listener,
             }))),
-            EvictionConfig::Lru(eviction_config) => Cache::Lru(Arc::new(RawCache::new(RawCacheConfig {
+            EvictionConfig::Lfu(eviction_config) => Cache::Lfu(Arc::new(GenericCache::new(GenericCacheConfig {
                 name: self.name,
                 capacity: self.capacity,
                 shards: self.shards,
                 eviction_config,
-                slab_initial_capacity: self.slab_initial_capacity,
-                slab_segment_size: self.slab_segment_size,
+                object_pool_capacity: self.object_pool_capacity,
                 hash_builder: self.hash_builder,
                 weighter: self.weighter,
                 event_listener: self.event_listener,
             }))),
-            EvictionConfig::Lfu(eviction_config) => Cache::Lfu(Arc::new(RawCache::new(RawCacheConfig {
+            EvictionConfig::S3Fifo(eviction_config) => Cache::S3Fifo(Arc::new(GenericCache::new(GenericCacheConfig {
                 name: self.name,
                 capacity: self.capacity,
                 shards: self.shards,
                 eviction_config,
-                slab_initial_capacity: self.slab_initial_capacity,
-                slab_segment_size: self.slab_segment_size,
+                object_pool_capacity: self.object_pool_capacity,
                 hash_builder: self.hash_builder,
                 weighter: self.weighter,
                 event_listener: self.event_listener,
@@ -476,9 +482,9 @@ where
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
             Self::Fifo(_) => f.debug_tuple("Cache::FifoCache").finish(),
-            Self::S3Fifo(_) => f.debug_tuple("Cache::S3FifoCache").finish(),
             Self::Lru(_) => f.debug_tuple("Cache::LruCache").finish(),
             Self::Lfu(_) => f.debug_tuple("Cache::LfuCache").finish(),
+            Self::S3Fifo(_) => f.debug_tuple("Cache::S3FifoCache").finish(),
         }
     }
 }
@@ -492,9 +498,9 @@ where
     fn clone(&self) -> Self {
         match self {
             Self::Fifo(cache) => Self::Fifo(cache.clone()),
-            Self::S3Fifo(cache) => Self::S3Fifo(cache.clone()),
             Self::Lru(cache) => Self::Lru(cache.clone()),
             Self::Lfu(cache) => Self::Lfu(cache.clone()),
+            Self::S3Fifo(cache) => Self::S3Fifo(cache.clone()),
         }
     }
 }
@@ -510,20 +516,20 @@ where
     pub fn insert(&self, key: K, value: V) -> CacheEntry<K, V, S> {
         match self {
             Cache::Fifo(cache) => cache.insert(key, value).into(),
-            Cache::S3Fifo(cache) => cache.insert(key, value).into(),
             Cache::Lru(cache) => cache.insert(key, value).into(),
             Cache::Lfu(cache) => cache.insert(key, value).into(),
+            Cache::S3Fifo(cache) => cache.insert(key, value).into(),
         }
     }
 
-    /// Insert cache entry with cache hint to the in-memory cache.
-    #[fastrace::trace(name = "foyer::memory::cache::insert_with_hint")]
-    pub fn insert_with_hint(&self, key: K, value: V, hint: CacheHint) -> CacheEntry<K, V, S> {
+    /// Insert cache entry with cache context to the in-memory cache.
+    #[fastrace::trace(name = "foyer::memory::cache::insert_with_context")]
+    pub fn insert_with_context(&self, key: K, value: V, context: CacheContext) -> CacheEntry<K, V, S> {
         match self {
-            Cache::Fifo(cache) => cache.insert_with_hint(key, value, hint.into()).into(),
-            Cache::S3Fifo(cache) => cache.insert_with_hint(key, value, hint.into()).into(),
-            Cache::Lru(cache) => cache.insert_with_hint(key, value, hint.into()).into(),
-            Cache::Lfu(cache) => cache.insert_with_hint(key, value, hint.into()).into(),
+            Cache::Fifo(cache) => cache.insert_with_context(key, value, context).into(),
+            Cache::Lru(cache) => cache.insert_with_context(key, value, context).into(),
+            Cache::Lfu(cache) => cache.insert_with_context(key, value, context).into(),
+            Cache::S3Fifo(cache) => cache.insert_with_context(key, value, context).into(),
         }
     }
 
@@ -532,28 +538,28 @@ where
     /// The entry will be removed as soon as the returned entry is dropped.
     ///
     /// The entry will become a normal entry after it is accessed.
-    #[fastrace::trace(name = "foyer::memory::cache::insert_ephemeral")]
-    pub fn insert_ephemeral(&self, key: K, value: V) -> CacheEntry<K, V, S> {
+    #[fastrace::trace(name = "foyer::memory::cache::deposit")]
+    pub fn deposit(&self, key: K, value: V) -> CacheEntry<K, V, S> {
         match self {
-            Cache::Fifo(cache) => cache.insert_ephemeral(key, value).into(),
-            Cache::S3Fifo(cache) => cache.insert_ephemeral(key, value).into(),
-            Cache::Lru(cache) => cache.insert_ephemeral(key, value).into(),
-            Cache::Lfu(cache) => cache.insert_ephemeral(key, value).into(),
+            Cache::Fifo(cache) => cache.deposit(key, value).into(),
+            Cache::Lru(cache) => cache.deposit(key, value).into(),
+            Cache::Lfu(cache) => cache.deposit(key, value).into(),
+            Cache::S3Fifo(cache) => cache.deposit(key, value).into(),
         }
     }
 
-    /// Temporarily insert cache entry with cache hint to the in-memory cache.
+    /// Temporarily insert cache entry with cache context to the in-memory cache.
     ///
     /// The entry will be removed as soon as the returned entry is dropped.
     ///
     /// The entry will become a normal entry after it is accessed.
-    #[fastrace::trace(name = "foyer::memory::cache::insert_ephemeral_with_hint")]
-    pub fn insert_ephemeral_with_hint(&self, key: K, value: V, hint: CacheHint) -> CacheEntry<K, V, S> {
+    #[fastrace::trace(name = "foyer::memory::cache::deposit_with_context")]
+    pub fn deposit_with_context(&self, key: K, value: V, context: CacheContext) -> CacheEntry<K, V, S> {
         match self {
-            Cache::Fifo(cache) => cache.insert_ephemeral_with_hint(key, value, hint.into()).into(),
-            Cache::Lru(cache) => cache.insert_ephemeral_with_hint(key, value, hint.into()).into(),
-            Cache::Lfu(cache) => cache.insert_ephemeral_with_hint(key, value, hint.into()).into(),
-            Cache::S3Fifo(cache) => cache.insert_ephemeral_with_hint(key, value, hint.into()).into(),
+            Cache::Fifo(cache) => cache.deposit_with_context(key, value, context).into(),
+            Cache::Lru(cache) => cache.deposit_with_context(key, value, context).into(),
+            Cache::Lfu(cache) => cache.deposit_with_context(key, value, context).into(),
+            Cache::S3Fifo(cache) => cache.deposit_with_context(key, value, context).into(),
         }
     }
 
@@ -565,9 +571,9 @@ where
     {
         match self {
             Cache::Fifo(cache) => cache.remove(key).map(CacheEntry::from),
-            Cache::S3Fifo(cache) => cache.remove(key).map(CacheEntry::from),
             Cache::Lru(cache) => cache.remove(key).map(CacheEntry::from),
             Cache::Lfu(cache) => cache.remove(key).map(CacheEntry::from),
+            Cache::S3Fifo(cache) => cache.remove(key).map(CacheEntry::from),
         }
     }
 
@@ -579,9 +585,9 @@ where
     {
         match self {
             Cache::Fifo(cache) => cache.get(key).map(CacheEntry::from),
-            Cache::S3Fifo(cache) => cache.get(key).map(CacheEntry::from),
             Cache::Lru(cache) => cache.get(key).map(CacheEntry::from),
             Cache::Lfu(cache) => cache.get(key).map(CacheEntry::from),
+            Cache::S3Fifo(cache) => cache.get(key).map(CacheEntry::from),
         }
     }
 
@@ -593,9 +599,9 @@ where
     {
         match self {
             Cache::Fifo(cache) => cache.contains(key),
-            Cache::S3Fifo(cache) => cache.contains(key),
             Cache::Lru(cache) => cache.contains(key),
             Cache::Lfu(cache) => cache.contains(key),
+            Cache::S3Fifo(cache) => cache.contains(key),
         }
     }
 
@@ -609,9 +615,9 @@ where
     {
         match self {
             Cache::Fifo(cache) => cache.touch(key),
-            Cache::S3Fifo(cache) => cache.touch(key),
             Cache::Lru(cache) => cache.touch(key),
             Cache::Lfu(cache) => cache.touch(key),
+            Cache::S3Fifo(cache) => cache.touch(key),
         }
     }
 
@@ -620,9 +626,9 @@ where
     pub fn clear(&self) {
         match self {
             Cache::Fifo(cache) => cache.clear(),
-            Cache::S3Fifo(cache) => cache.clear(),
             Cache::Lru(cache) => cache.clear(),
             Cache::Lfu(cache) => cache.clear(),
+            Cache::S3Fifo(cache) => cache.clear(),
         }
     }
 
@@ -630,9 +636,9 @@ where
     pub fn capacity(&self) -> usize {
         match self {
             Cache::Fifo(cache) => cache.capacity(),
-            Cache::S3Fifo(cache) => cache.capacity(),
             Cache::Lru(cache) => cache.capacity(),
             Cache::Lfu(cache) => cache.capacity(),
+            Cache::S3Fifo(cache) => cache.capacity(),
         }
     }
 
@@ -640,9 +646,9 @@ where
     pub fn usage(&self) -> usize {
         match self {
             Cache::Fifo(cache) => cache.usage(),
-            Cache::S3Fifo(cache) => cache.usage(),
             Cache::Lru(cache) => cache.usage(),
             Cache::Lfu(cache) => cache.usage(),
+            Cache::S3Fifo(cache) => cache.usage(),
         }
     }
 
@@ -658,9 +664,9 @@ where
     pub fn hash_builder(&self) -> &S {
         match self {
             Cache::Fifo(cache) => cache.hash_builder(),
-            Cache::S3Fifo(cache) => cache.hash_builder(),
             Cache::Lru(cache) => cache.hash_builder(),
             Cache::Lfu(cache) => cache.hash_builder(),
+            Cache::S3Fifo(cache) => cache.hash_builder(),
         }
     }
 
@@ -668,9 +674,9 @@ where
     pub fn shards(&self) -> usize {
         match self {
             Cache::Fifo(cache) => cache.shards(),
-            Cache::S3Fifo(cache) => cache.shards(),
             Cache::Lru(cache) => cache.shards(),
             Cache::Lfu(cache) => cache.shards(),
+            Cache::S3Fifo(cache) => cache.shards(),
         }
     }
 }
@@ -685,12 +691,12 @@ where
 {
     /// A future that is used to get entry value from the remote storage for the in-memory FIFO cache.
     Fifo(#[pin] FifoFetch<K, V, ER, S>),
-    /// A future that is used to get entry value from the remote storage for the in-memory S3FIFO cache.
-    S3Fifo(#[pin] S3FifoFetch<K, V, ER, S>),
     /// A future that is used to get entry value from the remote storage for the in-memory LRU cache.
     Lru(#[pin] LruFetch<K, V, ER, S>),
     /// A future that is used to get entry value from the remote storage for the in-memory LFU cache.
     Lfu(#[pin] LfuFetch<K, V, ER, S>),
+    /// A future that is used to get entry value from the remote storage for the in-memory S3FIFO cache.
+    S3Fifo(#[pin] S3FifoFetch<K, V, ER, S>),
 }
 
 impl<K, V, ER, S> From<FifoFetch<K, V, ER, S>> for Fetch<K, V, ER, S>
@@ -701,17 +707,6 @@ where
 {
     fn from(entry: FifoFetch<K, V, ER, S>) -> Self {
         Self::Fifo(entry)
-    }
-}
-
-impl<K, V, ER, S> From<S3FifoFetch<K, V, ER, S>> for Fetch<K, V, ER, S>
-where
-    K: Key,
-    V: Value,
-    S: HashBuilder,
-{
-    fn from(entry: S3FifoFetch<K, V, ER, S>) -> Self {
-        Self::S3Fifo(entry)
     }
 }
 
@@ -737,6 +732,17 @@ where
     }
 }
 
+impl<K, V, ER, S> From<S3FifoFetch<K, V, ER, S>> for Fetch<K, V, ER, S>
+where
+    K: Key,
+    V: Value,
+    S: HashBuilder,
+{
+    fn from(entry: S3FifoFetch<K, V, ER, S>) -> Self {
+        Self::S3Fifo(entry)
+    }
+}
+
 impl<K, V, ER, S> Future for Fetch<K, V, ER, S>
 where
     K: Key,
@@ -749,9 +755,9 @@ where
     fn poll(self: std::pin::Pin<&mut Self>, cx: &mut std::task::Context<'_>) -> std::task::Poll<Self::Output> {
         match self.project() {
             FetchProj::Fifo(entry) => entry.poll(cx).map(|res| res.map(CacheEntry::from)),
-            FetchProj::S3Fifo(entry) => entry.poll(cx).map(|res| res.map(CacheEntry::from)),
             FetchProj::Lru(entry) => entry.poll(cx).map(|res| res.map(CacheEntry::from)),
             FetchProj::Lfu(entry) => entry.poll(cx).map(|res| res.map(CacheEntry::from)),
+            FetchProj::S3Fifo(entry) => entry.poll(cx).map(|res| res.map(CacheEntry::from)),
         }
     }
 }
@@ -766,9 +772,9 @@ where
     pub fn state(&self) -> FetchState {
         match self {
             Fetch::Fifo(fetch) => fetch.state(),
-            Fetch::S3Fifo(fetch) => fetch.state(),
             Fetch::Lru(fetch) => fetch.state(),
             Fetch::Lfu(fetch) => fetch.state(),
+            Fetch::S3Fifo(fetch) => fetch.state(),
         }
     }
 
@@ -777,9 +783,9 @@ where
     pub fn store(&self) -> &Option<FetchMark> {
         match self {
             Fetch::Fifo(fetch) => fetch.store(),
-            Fetch::S3Fifo(fetch) => fetch.store(),
             Fetch::Lru(fetch) => fetch.store(),
             Fetch::Lfu(fetch) => fetch.store(),
+            Fetch::S3Fifo(fetch) => fetch.store(),
         }
     }
 }
@@ -804,29 +810,29 @@ where
     {
         match self {
             Cache::Fifo(cache) => Fetch::from(cache.fetch(key, fetch)),
-            Cache::S3Fifo(cache) => Fetch::from(cache.fetch(key, fetch)),
             Cache::Lru(cache) => Fetch::from(cache.fetch(key, fetch)),
             Cache::Lfu(cache) => Fetch::from(cache.fetch(key, fetch)),
+            Cache::S3Fifo(cache) => Fetch::from(cache.fetch(key, fetch)),
         }
     }
 
-    /// Get the cached entry with the given key and hint from the in-memory cache.
+    /// Get the cached entry with the given key and context from the in-memory cache.
     ///
     /// Use `fetch` to fetch the cache value from the remote storage on cache miss.
     ///
     /// The concurrent fetch requests will be deduplicated.
-    #[fastrace::trace(name = "foyer::memory::cache::fetch_with_hint")]
-    pub fn fetch_with_hint<F, FU, ER>(&self, key: K, hint: CacheHint, fetch: F) -> Fetch<K, V, ER, S>
+    #[fastrace::trace(name = "foyer::memory::cache::fetch_with_context")]
+    pub fn fetch_with_context<F, FU, ER>(&self, key: K, context: CacheContext, fetch: F) -> Fetch<K, V, ER, S>
     where
         F: FnOnce() -> FU,
         FU: Future<Output = std::result::Result<V, ER>> + Send + 'static,
         ER: Send + 'static + Debug,
     {
         match self {
-            Cache::Fifo(cache) => Fetch::from(cache.fetch_with_hint(key, hint.into(), fetch)),
-            Cache::S3Fifo(cache) => Fetch::from(cache.fetch_with_hint(key, hint.into(), fetch)),
-            Cache::Lru(cache) => Fetch::from(cache.fetch_with_hint(key, hint.into(), fetch)),
-            Cache::Lfu(cache) => Fetch::from(cache.fetch_with_hint(key, hint.into(), fetch)),
+            Cache::Fifo(cache) => Fetch::from(cache.fetch_with_context(key, context, fetch)),
+            Cache::Lru(cache) => Fetch::from(cache.fetch_with_context(key, context, fetch)),
+            Cache::Lfu(cache) => Fetch::from(cache.fetch_with_context(key, context, fetch)),
+            Cache::S3Fifo(cache) => Fetch::from(cache.fetch_with_context(key, context, fetch)),
         }
     }
 
@@ -841,7 +847,7 @@ where
     pub fn fetch_inner<F, FU, ER, ID>(
         &self,
         key: K,
-        hint: CacheHint,
+        context: CacheContext,
         fetch: F,
         runtime: &SingletonHandle,
     ) -> Fetch<K, V, ER, S>
@@ -852,10 +858,10 @@ where
         ID: Into<Diversion<std::result::Result<V, ER>, FetchMark>>,
     {
         match self {
-            Cache::Fifo(cache) => Fetch::from(cache.fetch_inner(key, hint.into(), fetch, runtime)),
-            Cache::Lru(cache) => Fetch::from(cache.fetch_inner(key, hint.into(), fetch, runtime)),
-            Cache::Lfu(cache) => Fetch::from(cache.fetch_inner(key, hint.into(), fetch, runtime)),
-            Cache::S3Fifo(cache) => Fetch::from(cache.fetch_inner(key, hint.into(), fetch, runtime)),
+            Cache::Fifo(cache) => Fetch::from(cache.fetch_inner(key, context, fetch, runtime)),
+            Cache::Lru(cache) => Fetch::from(cache.fetch_inner(key, context, fetch, runtime)),
+            Cache::Lfu(cache) => Fetch::from(cache.fetch_inner(key, context, fetch, runtime)),
+            Cache::S3Fifo(cache) => Fetch::from(cache.fetch_inner(key, context, fetch, runtime)),
         }
     }
 }
@@ -869,10 +875,11 @@ mod tests {
     use rand::{rngs::StdRng, seq::SliceRandom, Rng, SeedableRng};
 
     use super::*;
-    use crate::eviction::{fifo::FifoConfig, lfu::LfuConfig, lru::LruConfig, s3fifo::S3FifoConfig};
+    use crate::{eviction::s3fifo::S3FifoConfig, FifoConfig, LfuConfig, LruConfig};
 
     const CAPACITY: usize = 100;
     const SHARDS: usize = 4;
+    const OBJECT_POOL_CAPACITY: usize = 64;
     const RANGE: Range<u64> = 0..1000;
     const OPS: usize = 10000;
     const CONCURRENCY: usize = 8;
@@ -881,6 +888,7 @@ mod tests {
         CacheBuilder::new(CAPACITY)
             .with_shards(SHARDS)
             .with_eviction_config(FifoConfig {})
+            .with_object_pool_capacity(OBJECT_POOL_CAPACITY)
             .build()
     }
 
@@ -890,6 +898,7 @@ mod tests {
             .with_eviction_config(LruConfig {
                 high_priority_pool_ratio: 0.1,
             })
+            .with_object_pool_capacity(OBJECT_POOL_CAPACITY)
             .build()
     }
 
@@ -902,6 +911,7 @@ mod tests {
                 cmsketch_eps: 0.001,
                 cmsketch_confidence: 0.9,
             })
+            .with_object_pool_capacity(OBJECT_POOL_CAPACITY)
             .build()
     }
 
@@ -913,6 +923,7 @@ mod tests {
                 ghost_queue_capacity_ratio: 10.0,
                 small_to_main_freq_threshold: 2,
             })
+            .with_object_pool_capacity(OBJECT_POOL_CAPACITY)
             .build()
     }
 
@@ -997,7 +1008,7 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_cache_with_empty_initial_slab() {
-        case(CacheBuilder::new(8).with_slab_initial_capacity(0).build()).await
+    async fn test_cache_with_zero_object_pool() {
+        case(CacheBuilder::new(8).with_object_pool_capacity(0).build()).await
     }
 }
