+++ foyer-memory-bakcup/src/prelude.rs
@@ -12,15 +12,11 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-//! TODO(MrCroxx): Finish the crate level docs.
+pub use ahash::RandomState;
 
-mod cache;
-mod eviction;
-mod indexer;
-mod raw;
-mod record;
-mod slab;
-mod sync;
-
-mod prelude;
-pub use prelude::*;
+pub use crate::{
+    cache::{Cache, CacheBuilder, CacheEntry, EvictionConfig, Fetch},
+    context::CacheContext,
+    eviction::{fifo::FifoConfig, lfu::LfuConfig, lru::LruConfig, s3fifo::S3FifoConfig},
+    generic::{FetchMark, FetchState, Weighter},
+};
