+++ foyer-memory/Cargo.toml
@@ -30,29 +30,22 @@ tokio = { workspace = true }
 tracing = "0.1"
 
 [dev-dependencies]
-anyhow = "1"
-csv = "1.3.0"
-moka = { version = "0.12", features = ["sync"] }
+# anyhow = "1"
+# csv = "1.3.0"
+# moka = { version = "0.12", features = ["sync"] }
 rand = { version = "0.8", features = ["small_rng"] }
 test-log = { workspace = true }
-zipf = "7.0.1"
+# zipf = "7.0.1"
 
 [features]
 deadlock = ["parking_lot/deadlock_detection"]
+# FIXME: remove sanity feature
+sanity = []
 strict_assertions = [
     "foyer-common/strict_assertions",
     "foyer-intrusive/strict_assertions",
 ]
-sanity = ["strict_assertions"]
 tracing = ["fastrace/enable", "foyer-common/tracing"]
-
-[[bench]]
-name = "bench_hit_ratio"
-harness = false
-
-[[bench]]
-name = "bench_dynamic_dispatch"
-harness = false
 
 [lints]
 workspace = true
