+++ foyer-memory/src/eviction/lfu.rs
@@ -12,10 +12,14 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-use std::{fmt::Debug, ptr::NonNull};
+use std::ptr::NonNull;
 
 use cmsketch::CMSketchU16;
-use foyer_common::{assert::OptionExt, strict_assert, strict_assert_eq, strict_assert_ne};
+use foyer_common::{
+    assert::OptionExt,
+    code::{Key, Value},
+    strict_assert, strict_assert_eq, strict_assert_ne,
+};
 use foyer_intrusive::{
     adapter::Link,
     dlist::{Dlist, DlistLink},
@@ -23,11 +27,8 @@ use foyer_intrusive::{
 };
 use serde::{Deserialize, Serialize};
 
-use crate::{
-    eviction::Eviction,
-    handle::{BaseHandle, Handle},
-    CacheContext,
-};
+use super::{Eviction, Operator};
+use crate::record::{CacheHint, Record};
 
 /// w-TinyLFU eviction algorithm config.
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -67,18 +68,19 @@ impl Default for LfuConfig {
     }
 }
 
-#[derive(Debug, Clone)]
-pub struct LfuContext(CacheContext);
+/// w-TinyLFU eviction algorithm hint.
+#[derive(Debug, Clone, Default)]
+pub struct LfuHint;
 
-impl From<CacheContext> for LfuContext {
-    fn from(context: CacheContext) -> Self {
-        Self(context)
+impl From<CacheHint> for LfuHint {
+    fn from(_: CacheHint) -> Self {
+        LfuHint
     }
 }
 
-impl From<LfuContext> for CacheContext {
-    fn from(context: LfuContext) -> Self {
-        context.0
+impl From<LfuHint> for CacheHint {
+    fn from(_: LfuHint) -> Self {
+        CacheHint::Normal
     }
 }
 
@@ -90,57 +92,20 @@ enum Queue {
     Protected,
 }
 
-pub struct LfuHandle<T>
-where
-    T: Send + Sync + 'static,
-{
+impl Default for Queue {
+    fn default() -> Self {
+        Self::None
+    }
+}
+
+/// w-TinyLFU eviction algorithm hint.
+#[derive(Debug, Default)]
+pub struct LfuState {
     link: DlistLink,
-    base: BaseHandle<T, LfuContext>,
     queue: Queue,
 }
 
-impl<T> Debug for LfuHandle<T>
-where
-    T: Send + Sync + 'static,
-{
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("LfuHandle").finish()
-    }
-}
-
-intrusive_adapter! { LfuHandleDlistAdapter<T> = LfuHandle<T> { link: DlistLink } where T: Send + Sync + 'static }
-
-impl<T> Default for LfuHandle<T>
-where
-    T: Send + Sync + 'static,
-{
-    fn default() -> Self {
-        Self {
-            link: DlistLink::default(),
-            base: BaseHandle::new(),
-            queue: Queue::None,
-        }
-    }
-}
-
-impl<T> Handle for LfuHandle<T>
-where
-    T: Send + Sync + 'static,
-{
-    type Data = T;
-    type Context = LfuContext;
-
-    fn base(&self) -> &BaseHandle<Self::Data, Self::Context> {
-        &self.base
-    }
-
-    fn base_mut(&mut self) -> &mut BaseHandle<Self::Data, Self::Context> {
-        &mut self.base
-    }
-}
-
-unsafe impl<T> Send for LfuHandle<T> where T: Send + Sync + 'static {}
-unsafe impl<T> Sync for LfuHandle<T> where T: Send + Sync + 'static {}
+intrusive_adapter! { Adapter<K, V> = Record<Lfu<K, V>> { state.link => DlistLink } where K: Key, V: Value }
 
 /// This implementation is inspired by [Caffeine](https://github.com/ben-manes/caffeine) under Apache License 2.0
 ///
@@ -154,13 +119,14 @@ unsafe impl<T> Sync for LfuHandle<T> where T: Send + Sync + 'static {}
 ///
 /// When evicting, the entry with a lower frequency from `window` or `probation` will be evicted first, then from
 /// `protected`.
-pub struct Lfu<T>
+pub struct Lfu<K, V>
 where
-    T: Send + Sync + 'static,
+    K: Key,
+    V: Value,
 {
-    window: Dlist<LfuHandleDlistAdapter<T>>,
-    probation: Dlist<LfuHandleDlistAdapter<T>>,
-    protected: Dlist<LfuHandleDlistAdapter<T>>,
+    window: Dlist<Adapter<K, V>>,
+    probation: Dlist<Adapter<K, V>>,
+    protected: Dlist<Adapter<K, V>>,
 
     window_weight: usize,
     probation_weight: usize,
@@ -169,19 +135,21 @@ where
     window_weight_capacity: usize,
     protected_weight_capacity: usize,
 
+    // TODO(MrCroxx): use a count-min-sketch impl with atomic u16
     frequencies: CMSketchU16,
 
     step: usize,
     decay: usize,
 }
 
-impl<T> Lfu<T>
+impl<K, V> Lfu<K, V>
 where
-    T: Send + Sync + 'static,
+    K: Key,
+    V: Value,
 {
-    fn increase_queue_weight(&mut self, handle: &LfuHandle<T>) {
-        let weight = handle.base().weight();
-        match handle.queue {
+    fn increase_queue_weight(&mut self, record: &Record<Lfu<K, V>>) {
+        let weight = record.weight();
+        match record.state.queue {
             Queue::None => unreachable!(),
             Queue::Window => self.window_weight += weight,
             Queue::Probation => self.probation_weight += weight,
@@ -189,9 +157,9 @@ where
         }
     }
 
-    fn decrease_queue_weight(&mut self, handle: &LfuHandle<T>) {
-        let weight = handle.base().weight();
-        match handle.queue {
+    fn decrease_queue_weight(&mut self, record: &Record<Lfu<K, V>>) {
+        let weight = record.weight();
+        match record.state.queue {
             Queue::None => unreachable!(),
             Queue::Window => self.window_weight -= weight,
             Queue::Probation => self.probation_weight -= weight,
@@ -209,14 +177,18 @@ where
     }
 }
 
-impl<T> Eviction for Lfu<T>
+impl<K, V> Eviction for Lfu<K, V>
 where
-    T: Send + Sync + 'static,
+    K: Key,
+    V: Value,
 {
-    type Handle = LfuHandle<T>;
     type Config = LfuConfig;
+    type Key = K;
+    type Value = V;
+    type Hint = LfuHint;
+    type State = LfuState;
 
-    unsafe fn new(capacity: usize, config: &Self::Config) -> Self
+    fn new(capacity: usize, config: &Self::Config) -> Self
     where
         Self: Sized,
     {
@@ -258,33 +230,64 @@ where
         }
     }
 
-    unsafe fn push(&mut self, mut ptr: NonNull<Self::Handle>) {
-        let handle = ptr.as_mut();
+    fn update(&mut self, capacity: usize, config: &Self::Config) {
+        if config.window_capacity_ratio <= 0.0 || config.window_capacity_ratio >= 1.0 {
+            tracing::error!(
+                "window_capacity_ratio must be in (0, 1), given: {}, new config ignored",
+                config.window_capacity_ratio
+            );
+        }
 
-        strict_assert!(!handle.link.is_linked());
-        strict_assert!(!handle.base().is_in_eviction());
-        strict_assert_eq!(handle.queue, Queue::None);
+        if config.protected_capacity_ratio <= 0.0 || config.protected_capacity_ratio >= 1.0 {
+            tracing::error!(
+                "protected_capacity_ratio must be in (0, 1), given: {}, new config ignored",
+                config.protected_capacity_ratio
+            );
+        }
+
+        if config.window_capacity_ratio + config.protected_capacity_ratio >= 1.0 {
+            tracing::error!(
+                "must guarantee: window_capacity_ratio + protected_capacity_ratio < 1, given: {}, new config ignored",
+                config.window_capacity_ratio + config.protected_capacity_ratio
+            )
+        }
+
+        // TODO(MrCroxx): Raise a warn log the cmsketch args updates is not supported yet if it is modified.
+
+        let window_weight_capacity = (capacity as f64 * config.window_capacity_ratio) as usize;
+        let protected_weight_capacity = (capacity as f64 * config.protected_capacity_ratio) as usize;
+
+        self.window_weight_capacity = window_weight_capacity;
+        self.protected_weight_capacity = protected_weight_capacity;
+    }
+
+    fn push(&mut self, mut ptr: NonNull<Record<Self>>) {
+        let record = unsafe { ptr.as_mut() };
+
+        strict_assert!(!record.state.link.is_linked());
+        strict_assert!(!record.is_in_eviction());
+        strict_assert_eq!(record.state.queue, Queue::None);
 
         self.window.push_back(ptr);
-        handle.base_mut().set_in_eviction(true);
-        handle.queue = Queue::Window;
+        record.set_in_eviction(true);
+        record.state.queue = Queue::Window;
 
-        self.increase_queue_weight(handle);
-        self.update_frequencies(handle.base().hash());
+        self.increase_queue_weight(record);
+        self.update_frequencies(record.hash());
 
         // If `window` weight exceeds the capacity, overflow entry from `window` to `probation`.
         while self.window_weight > self.window_weight_capacity {
             strict_assert!(!self.window.is_empty());
-            let mut ptr = self.window.pop_front().strict_unwrap_unchecked();
-            let handle = ptr.as_mut();
-            self.decrease_queue_weight(handle);
-            handle.queue = Queue::Probation;
-            self.increase_queue_weight(handle);
-            self.probation.push_back(ptr);
+            let mut p = unsafe { self.window.pop_front().strict_unwrap_unchecked() };
+            let r = unsafe { p.as_mut() };
+            self.decrease_queue_weight(r);
+            r.state.queue = Queue::Probation;
+            self.increase_queue_weight(r);
+            self.probation.push_back(p);
         }
     }
 
-    unsafe fn pop(&mut self) -> Option<NonNull<Self::Handle>> {
+    fn pop(&mut self) -> Option<NonNull<Record<Self>>> {
         // Compare the frequency of the front element of `window` and `probation` queue, and evict the lower one.
         // If both `window` and `probation` are empty, try evict from `protected`.
         let mut ptr = match (self.window.front(), self.probation.front()) {
@@ -292,8 +295,7 @@ where
             (None, Some(_)) => self.probation.pop_front(),
             (Some(_), None) => self.window.pop_front(),
             (Some(window), Some(probation)) => {
-                if self.frequencies.estimate(window.base().hash()) < self.frequencies.estimate(probation.base().hash())
-                {
+                if self.frequencies.estimate(window.hash()) < self.frequencies.estimate(probation.hash()) {
                     self.window.pop_front()
 
                     // TODO(MrCroxx): Rotate probation to prevent a high frequency but cold head holds back promotion
@@ -305,118 +307,114 @@ where
         }
         .or_else(|| self.protected.pop_front())?;
 
-        let handle = ptr.as_mut();
+        let record = unsafe { ptr.as_mut() };
 
-        strict_assert!(!handle.link.is_linked());
-        strict_assert!(handle.base().is_in_eviction());
-        strict_assert_ne!(handle.queue, Queue::None);
+        strict_assert!(!record.state.link.is_linked());
+        strict_assert!(record.is_in_eviction());
+        strict_assert_ne!(record.state.queue, Queue::None);
 
-        self.decrease_queue_weight(handle);
-        handle.queue = Queue::None;
-        handle.base_mut().set_in_eviction(false);
+        self.decrease_queue_weight(record);
+        record.state.queue = Queue::None;
+        record.set_in_eviction(false);
 
         Some(ptr)
     }
 
-    unsafe fn release(&mut self, mut ptr: NonNull<Self::Handle>) {
-        let handle = ptr.as_mut();
+    fn remove(&mut self, mut ptr: NonNull<Record<Self>>) {
+        let record = unsafe { ptr.as_mut() };
 
-        match handle.queue {
-            Queue::None => {
-                strict_assert!(!handle.link.is_linked());
-                strict_assert!(!handle.base().is_in_eviction());
-                self.push(ptr);
-                strict_assert!(handle.link.is_linked());
-                strict_assert!(handle.base().is_in_eviction());
-            }
-            Queue::Window => {
-                // Move to MRU position of `window`.
-                strict_assert!(handle.link.is_linked());
-                strict_assert!(handle.base().is_in_eviction());
-                self.window.remove_raw(handle.link.raw());
-                self.window.push_back(ptr);
-            }
-            Queue::Probation => {
-                // Promote to MRU position of `protected`.
-                strict_assert!(handle.link.is_linked());
-                strict_assert!(handle.base().is_in_eviction());
-                self.probation.remove_raw(handle.link.raw());
-                self.decrease_queue_weight(handle);
-                handle.queue = Queue::Protected;
-                self.increase_queue_weight(handle);
-                self.protected.push_back(ptr);
+        strict_assert!(record.state.link.is_linked());
+        strict_assert!(record.is_in_eviction());
+        strict_assert_ne!(record.state.queue, Queue::None);
 
-                // If `protected` weight exceeds the capacity, overflow entry from `protected` to `probation`.
-                while self.protected_weight > self.protected_weight_capacity {
-                    strict_assert!(!self.protected.is_empty());
-                    let mut ptr = self.protected.pop_front().strict_unwrap_unchecked();
-                    let handle = ptr.as_mut();
-                    self.decrease_queue_weight(handle);
-                    handle.queue = Queue::Probation;
-                    self.increase_queue_weight(handle);
-                    self.probation.push_back(ptr);
-                }
-            }
-            Queue::Protected => {
-                // Move to MRU position of `protected`.
-                strict_assert!(handle.link.is_linked());
-                strict_assert!(handle.base().is_in_eviction());
-                self.protected.remove_raw(handle.link.raw());
-                self.protected.push_back(ptr);
-            }
-        }
-    }
-
-    unsafe fn acquire(&mut self, ptr: NonNull<Self::Handle>) {
-        self.update_frequencies(ptr.as_ref().base().hash());
-    }
-
-    unsafe fn remove(&mut self, mut ptr: NonNull<Self::Handle>) {
-        let handle = ptr.as_mut();
-
-        strict_assert!(handle.link.is_linked());
-        strict_assert!(handle.base().is_in_eviction());
-        strict_assert_ne!(handle.queue, Queue::None);
-
-        match handle.queue {
+        match record.state.queue {
             Queue::None => unreachable!(),
-            Queue::Window => self.window.remove_raw(handle.link.raw()),
-            Queue::Probation => self.probation.remove_raw(handle.link.raw()),
-            Queue::Protected => self.protected.remove_raw(handle.link.raw()),
+            Queue::Window => self.window.remove(ptr),
+            Queue::Probation => self.probation.remove(ptr),
+            Queue::Protected => self.protected.remove(ptr),
         };
 
-        strict_assert!(!handle.link.is_linked());
+        strict_assert!(!record.state.link.is_linked());
 
-        self.decrease_queue_weight(handle);
-        handle.queue = Queue::None;
-        handle.base_mut().set_in_eviction(false);
+        self.decrease_queue_weight(record);
+        record.state.queue = Queue::None;
+        record.set_in_eviction(false);
     }
 
-    unsafe fn clear(&mut self) -> Vec<NonNull<Self::Handle>> {
-        let mut res = Vec::with_capacity(self.len());
-
-        while !self.is_empty() {
-            let ptr = self.pop().strict_unwrap_unchecked();
-            strict_assert!(!ptr.as_ref().base().is_in_eviction());
-            strict_assert!(!ptr.as_ref().link.is_linked());
-            strict_assert_eq!(ptr.as_ref().queue, Queue::None);
-            res.push(ptr);
+    fn clear(&mut self) {
+        while let Some(ptr) = self.pop() {
+            strict_assert!(!unsafe { ptr.as_ref() }.is_in_eviction());
+            strict_assert!(!unsafe { ptr.as_ref() }.state.link.is_linked());
+            strict_assert_eq!(unsafe { ptr.as_ref() }.state.queue, Queue::None);
         }
-
-        res
     }
 
     fn len(&self) -> usize {
         self.window.len() + self.probation.len() + self.protected.len()
     }
 
-    fn is_empty(&self) -> bool {
-        self.len() == 0
+    fn acquire_operator() -> super::Operator {
+        // TODO(MrCroxx): use a count-min-sketch with atomic u16 impl.
+        Operator::Mutable
+    }
+
+    fn acquire_immutable(&self, _ptr: NonNull<Record<Self>>) {
+        unreachable!()
+    }
+
+    fn acquire_mutable(&mut self, ptr: NonNull<Record<Self>>) {
+        self.update_frequencies(unsafe { ptr.as_ref() }.hash());
+    }
+
+    fn release(&mut self, mut ptr: NonNull<Record<Self>>) {
+        let record = unsafe { ptr.as_mut() };
+
+        match record.state.queue {
+            Queue::None => {
+                strict_assert!(!record.state.link.is_linked());
+                strict_assert!(!record.is_in_eviction());
+                self.push(ptr);
+                strict_assert!(record.state.link.is_linked());
+                strict_assert!(record.is_in_eviction());
+            }
+            Queue::Window => {
+                // Move to MRU position of `window`.
+                strict_assert!(record.state.link.is_linked());
+                strict_assert!(record.is_in_eviction());
+                self.window.remove(ptr);
+                self.window.push_back(ptr);
+            }
+            Queue::Probation => {
+                // Promote to MRU position of `protected`.
+                strict_assert!(record.state.link.is_linked());
+                strict_assert!(record.is_in_eviction());
+                self.probation.remove(ptr);
+                self.decrease_queue_weight(record);
+                record.state.queue = Queue::Protected;
+                self.increase_queue_weight(record);
+                self.protected.push_back(ptr);
+
+                // If `protected` weight exceeds the capacity, overflow entry from `protected` to `probation`.
+                while self.protected_weight > self.protected_weight_capacity {
+                    strict_assert!(!self.protected.is_empty());
+                    let mut p = unsafe { self.protected.pop_front().strict_unwrap_unchecked() };
+                    let r = unsafe { p.as_mut() };
+                    self.decrease_queue_weight(r);
+                    r.state.queue = Queue::Probation;
+                    self.increase_queue_weight(r);
+                    self.probation.push_back(p);
+                }
+            }
+            Queue::Protected => {
+                // Move to MRU position of `protected`.
+                strict_assert!(record.state.link.is_linked());
+                strict_assert!(record.is_in_eviction());
+                self.protected.remove(ptr);
+                self.protected.push_back(ptr);
+            }
+        }
     }
 }
-
-unsafe impl<T> Send for Lfu<T> where T: Send + Sync + 'static {}
-unsafe impl<T> Sync for Lfu<T> where T: Send + Sync + 'static {}
 
 #[cfg(test)]
 mod tests {
@@ -424,24 +422,23 @@ mod tests {
     use itertools::Itertools;
 
     use super::*;
-    use crate::{eviction::test_utils::TestEviction, handle::HandleExt};
+    use crate::{eviction::test_utils::TestEviction, record::Data};
 
-    impl<T> TestEviction for Lfu<T>
+    impl<K, V> TestEviction for Lfu<K, V>
     where
-        T: Send + Sync + 'static + Clone,
+        K: Key + Clone,
+        V: Value + Clone,
     {
-        fn dump(&self) -> Vec<T> {
+        fn dump(&self) -> Vec<NonNull<Record<Self>>> {
             self.window
-                .iter()
-                .chain(self.probation.iter())
-                .chain(self.protected.iter())
-                .map(|handle| handle.base().data_unwrap_unchecked().clone())
+                .iter_ptr()
+                .chain(self.probation.iter_ptr())
+                .chain(self.protected.iter_ptr())
                 .collect_vec()
         }
     }
 
-    type TestLfu = Lfu<u64>;
-    type TestLfuHandle = LfuHandle<u64>;
+    type TestLfu = Lfu<u64, u64>;
 
     unsafe fn assert_test_lfu(
         lfu: &TestLfu,
@@ -449,7 +446,7 @@ mod tests {
         window: usize,
         probation: usize,
         protected: usize,
-        entries: Vec<u64>,
+        entries: Vec<NonNull<Record<TestLfu>>>,
     ) {
         assert_eq!(lfu.len(), len);
         assert_eq!(lfu.window.len(), window);
@@ -472,8 +469,14 @@ mod tests {
         unsafe {
             let ptrs = (0..100)
                 .map(|i| {
-                    let mut handle = Box::<TestLfuHandle>::default();
-                    handle.init(i, i, 1, LfuContext(CacheContext::Default));
+                    let handle = Box::new(Record::new(Data {
+                        key: i,
+                        value: i,
+                        hint: LfuHint,
+                        state: Default::default(),
+                        hash: i,
+                        weight: 1,
+                    }));
                     NonNull::new_unchecked(Box::into_raw(handle))
                 })
                 .collect_vec();
@@ -487,19 +490,21 @@ mod tests {
             };
             let mut lfu = TestLfu::new(10, &config);
 
+            let ps = |indices: &[usize]| indices.iter().map(|&i| ptrs[i]).collect_vec();
+
             assert_eq!(lfu.window_weight_capacity, 2);
             assert_eq!(lfu.protected_weight_capacity, 6);
 
             lfu.push(ptrs[0]);
             lfu.push(ptrs[1]);
-            assert_test_lfu(&lfu, 2, 2, 0, 0, vec![0, 1]);
+            assert_test_lfu(&lfu, 2, 2, 0, 0, ps(&[0, 1]));
 
             lfu.push(ptrs[2]);
             lfu.push(ptrs[3]);
-            assert_test_lfu(&lfu, 4, 2, 2, 0, vec![2, 3, 0, 1]);
+            assert_test_lfu(&lfu, 4, 2, 2, 0, ps(&[2, 3, 0, 1]));
 
             (4..10).for_each(|i| lfu.push(ptrs[i]));
-            assert_test_lfu(&lfu, 10, 2, 8, 0, vec![8, 9, 0, 1, 2, 3, 4, 5, 6, 7]);
+            assert_test_lfu(&lfu, 10, 2, 8, 0, ps(&[8, 9, 0, 1, 2, 3, 4, 5, 6, 7]));
 
             (0..10).for_each(|i| assert_min_frequency(&lfu, i, 1));
 
@@ -509,35 +514,35 @@ mod tests {
 
             // [9, 0] [1, 2, 3, 4, 5, 6, 7, 8]
             lfu.release(p0);
-            assert_test_lfu(&lfu, 10, 2, 8, 0, vec![9, 0, 1, 2, 3, 4, 5, 6, 7, 8]);
+            assert_test_lfu(&lfu, 10, 2, 8, 0, ps(&[9, 0, 1, 2, 3, 4, 5, 6, 7, 8]));
 
             // [0, 9] [1, 2, 3, 4, 5, 6, 7, 8]
             lfu.release(ptrs[9]);
-            assert_test_lfu(&lfu, 10, 2, 8, 0, vec![0, 9, 1, 2, 3, 4, 5, 6, 7, 8]);
+            assert_test_lfu(&lfu, 10, 2, 8, 0, ps(&[0, 9, 1, 2, 3, 4, 5, 6, 7, 8]));
 
             // [0, 9] [1, 2, 7, 8] [3, 4, 5, 6]
             (3..7).for_each(|i| lfu.release(ptrs[i]));
-            assert_test_lfu(&lfu, 10, 2, 4, 4, vec![0, 9, 1, 2, 7, 8, 3, 4, 5, 6]);
+            assert_test_lfu(&lfu, 10, 2, 4, 4, ps(&[0, 9, 1, 2, 7, 8, 3, 4, 5, 6]));
 
             // [0, 9] [1, 2, 7, 8] [5, 6, 3, 4]
             (3..5).for_each(|i| lfu.release(ptrs[i]));
-            assert_test_lfu(&lfu, 10, 2, 4, 4, vec![0, 9, 1, 2, 7, 8, 5, 6, 3, 4]);
+            assert_test_lfu(&lfu, 10, 2, 4, 4, ps(&[0, 9, 1, 2, 7, 8, 5, 6, 3, 4]));
 
             // [0, 9] [5, 6] [3, 4, 1, 2, 7, 8]
             [1, 2, 7, 8].into_iter().for_each(|i| lfu.release(ptrs[i]));
-            assert_test_lfu(&lfu, 10, 2, 2, 6, vec![0, 9, 5, 6, 3, 4, 1, 2, 7, 8]);
+            assert_test_lfu(&lfu, 10, 2, 2, 6, ps(&[0, 9, 5, 6, 3, 4, 1, 2, 7, 8]));
 
             // [0, 9] [6] [3, 4, 1, 2, 7, 8]
             let p5 = lfu.pop().unwrap();
             assert_eq!(p5, ptrs[5]);
-            assert_test_lfu(&lfu, 9, 2, 1, 6, vec![0, 9, 6, 3, 4, 1, 2, 7, 8]);
+            assert_test_lfu(&lfu, 9, 2, 1, 6, ps(&[0, 9, 6, 3, 4, 1, 2, 7, 8]));
 
             (10..13).for_each(|i| lfu.push(ptrs[i]));
 
             // [11, 12] [6, 0, 9, 10] [3, 4, 1, 2, 7, 8]
-            assert_test_lfu(&lfu, 12, 2, 4, 6, vec![11, 12, 6, 0, 9, 10, 3, 4, 1, 2, 7, 8]);
+            assert_test_lfu(&lfu, 12, 2, 4, 6, ps(&[11, 12, 6, 0, 9, 10, 3, 4, 1, 2, 7, 8]));
             (1..13).for_each(|i| assert_min_frequency(&lfu, i, 0));
-            lfu.acquire(ptrs[0]);
+            lfu.acquire_mutable(ptrs[0]);
             assert_min_frequency(&lfu, 0, 2);
 
             // evict 11 because freq(11) < freq(0)
@@ -546,15 +551,10 @@ mod tests {
             let p11 = lfu.pop().unwrap();
             assert_eq!(p6, ptrs[6]);
             assert_eq!(p11, ptrs[11]);
-            assert_test_lfu(&lfu, 10, 1, 3, 6, vec![12, 0, 9, 10, 3, 4, 1, 2, 7, 8]);
+            assert_test_lfu(&lfu, 10, 1, 3, 6, ps(&[12, 0, 9, 10, 3, 4, 1, 2, 7, 8]));
 
-            assert_eq!(
-                lfu.clear(),
-                [12, 0, 9, 10, 3, 4, 1, 2, 7, 8]
-                    .into_iter()
-                    .map(|i| ptrs[i])
-                    .collect_vec()
-            );
+            lfu.clear();
+            assert_test_lfu(&lfu, 0, 0, 0, 0, vec![]);
 
             for ptr in ptrs {
                 let _ = Box::from_raw(ptr.as_ptr());
