+++ foyer-memory/src/eviction/lru.rs
@@ -12,9 +12,13 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-use std::{fmt::Debug, ptr::NonNull};
+use std::ptr::NonNull;
 
-use foyer_common::{assert::OptionExt, strict_assert, strict_assert_eq};
+use foyer_common::{
+    assert::OptionExt,
+    code::{Key, Value},
+    strict_assert, strict_assert_eq,
+};
 use foyer_intrusive::{
     adapter::Link,
     dlist::{Dlist, DlistLink},
@@ -22,13 +26,10 @@ use foyer_intrusive::{
 };
 use serde::{Deserialize, Serialize};
 
-use crate::{
-    eviction::Eviction,
-    handle::{BaseHandle, Handle},
-    CacheContext,
-};
+use super::{Eviction, Operator};
+use crate::record::{CacheHint, Record};
 
-/// LRU eviction algorithm config.
+/// Lru eviction algorithm config.
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct LruConfig {
     /// The ratio of the high priority pool occupied.
@@ -45,130 +46,101 @@ pub struct LruConfig {
 impl Default for LruConfig {
     fn default() -> Self {
         Self {
-            high_priority_pool_ratio: 0.0,
+            high_priority_pool_ratio: 0.9,
         }
     }
 }
 
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum LruContext {
+/// Lru eviction algorithm hint.
+#[derive(Debug, Clone)]
+pub enum LruHint {
     HighPriority,
     LowPriority,
 }
 
-impl From<CacheContext> for LruContext {
-    fn from(value: CacheContext) -> Self {
-        match value {
-            CacheContext::Default => Self::HighPriority,
-            CacheContext::LowPriority => Self::LowPriority,
+impl Default for LruHint {
+    fn default() -> Self {
+        Self::HighPriority
+    }
+}
+
+impl From<CacheHint> for LruHint {
+    fn from(hint: CacheHint) -> Self {
+        match hint {
+            CacheHint::Normal => LruHint::HighPriority,
+            CacheHint::Low => LruHint::LowPriority,
         }
     }
 }
 
-impl From<LruContext> for CacheContext {
-    fn from(value: LruContext) -> Self {
-        match value {
-            LruContext::HighPriority => CacheContext::Default,
-            LruContext::LowPriority => CacheContext::LowPriority,
+impl From<LruHint> for CacheHint {
+    fn from(hint: LruHint) -> Self {
+        match hint {
+            LruHint::HighPriority => CacheHint::Normal,
+            LruHint::LowPriority => CacheHint::Low,
         }
     }
 }
 
-pub struct LruHandle<T>
-where
-    T: Send + Sync + 'static,
-{
+/// Lru eviction algorithm state.
+#[derive(Debug, Default)]
+pub struct LruState {
     link: DlistLink,
-    base: BaseHandle<T, LruContext>,
     in_high_priority_pool: bool,
 }
 
-impl<T> Debug for LruHandle<T>
+intrusive_adapter! { Adapter<K, V> = Record<Lru<K, V>> { state.link => DlistLink } where K: Key, V: Value }
+
+pub struct Lru<K, V>
 where
-    T: Send + Sync + 'static,
+    K: Key,
+    V: Value,
 {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("LruHandle").finish()
-    }
-}
-
-intrusive_adapter! { LruHandleDlistAdapter<T> = LruHandle<T> { link: DlistLink } where T: Send + Sync + 'static }
-
-impl<T> Default for LruHandle<T>
-where
-    T: Send + Sync + 'static,
-{
-    fn default() -> Self {
-        Self {
-            link: DlistLink::default(),
-            base: BaseHandle::new(),
-            in_high_priority_pool: false,
-        }
-    }
-}
-
-impl<T> Handle for LruHandle<T>
-where
-    T: Send + Sync + 'static,
-{
-    type Data = T;
-    type Context = LruContext;
-
-    fn base(&self) -> &BaseHandle<Self::Data, Self::Context> {
-        &self.base
-    }
-
-    fn base_mut(&mut self) -> &mut BaseHandle<Self::Data, Self::Context> {
-        &mut self.base
-    }
-}
-
-unsafe impl<T> Send for LruHandle<T> where T: Send + Sync + 'static {}
-unsafe impl<T> Sync for LruHandle<T> where T: Send + Sync + 'static {}
-
-pub struct Lru<T>
-where
-    T: Send + Sync + 'static,
-{
-    high_priority_list: Dlist<LruHandleDlistAdapter<T>>,
-    list: Dlist<LruHandleDlistAdapter<T>>,
+    high_priority_list: Dlist<Adapter<K, V>>,
+    list: Dlist<Adapter<K, V>>,
 
     high_priority_weight: usize,
     high_priority_weight_capacity: usize,
 }
 
-impl<T> Lru<T>
+impl<K, V> Lru<K, V>
 where
-    T: Send + Sync + 'static,
+    K: Key,
+    V: Value,
 {
-    unsafe fn may_overflow_high_priority_pool(&mut self) {
+    fn may_overflow_high_priority_pool(&mut self) {
         while self.high_priority_weight > self.high_priority_weight_capacity {
             strict_assert!(!self.high_priority_list.is_empty());
 
             // overflow last entry in high priority pool to low priority pool
-            let mut ptr = self.high_priority_list.pop_front().strict_unwrap_unchecked();
-            strict_assert!(ptr.as_ref().in_high_priority_pool);
-            ptr.as_mut().in_high_priority_pool = false;
-            self.high_priority_weight -= ptr.as_ref().base().weight();
+            let mut ptr = unsafe { self.high_priority_list.pop_front().strict_unwrap_unchecked() };
+            let record = unsafe { ptr.as_mut() };
+            strict_assert!(record.state.in_high_priority_pool);
+            record.state.in_high_priority_pool = false;
+            self.high_priority_weight -= record.weight();
             self.list.push_back(ptr);
         }
     }
 }
 
-impl<T> Eviction for Lru<T>
+impl<K, V> Eviction for Lru<K, V>
 where
-    T: Send + Sync + 'static,
+    K: Key,
+    V: Value,
 {
-    type Handle = LruHandle<T>;
     type Config = LruConfig;
+    type Key = K;
+    type Value = V;
+    type Hint = LruHint;
+    type State = LruState;
 
-    unsafe fn new(capacity: usize, config: &Self::Config) -> Self
+    fn new(capacity: usize, config: &Self::Config) -> Self
     where
         Self: Sized,
     {
         assert!(
-            config.high_priority_pool_ratio >= 0.0 && config.high_priority_pool_ratio <= 1.0,
-            "high_priority_pool_ratio_percentage must be in [0, 100], given: {}",
+            (0.0..=1.0).contains(&config.high_priority_pool_ratio),
+            "high_priority_pool_ratio_percentage must be in 0.0..=1.0, given: {}",
             config.high_priority_pool_ratio
         );
 
@@ -182,111 +154,119 @@ where
         }
     }
 
-    unsafe fn push(&mut self, mut ptr: NonNull<Self::Handle>) {
-        let handle = ptr.as_mut();
+    fn update(&mut self, capacity: usize, config: &Self::Config) {
+        if !(0.0..=1.0).contains(&config.high_priority_pool_ratio) {
+            tracing::error!(
+                "[fifo]: high_priority_pool_ratio_percentage must be in 0.0..=1.0, given: {}, new configuration ignored",
+                config.high_priority_pool_ratio
+            );
+            return;
+        }
 
-        strict_assert!(!handle.link.is_linked());
+        let high_priority_weight_capacity = (capacity as f64 * config.high_priority_pool_ratio) as usize;
+        self.high_priority_weight_capacity = high_priority_weight_capacity;
 
-        match handle.base().context() {
-            LruContext::HighPriority => {
-                handle.in_high_priority_pool = true;
-                self.high_priority_weight += handle.base().weight();
+        self.may_overflow_high_priority_pool();
+    }
+
+    fn push(&mut self, mut ptr: NonNull<Record<Self>>) {
+        let record = unsafe { ptr.as_mut() };
+
+        strict_assert!(!record.state.link.is_linked());
+
+        match record.hint() {
+            LruHint::HighPriority => {
+                record.state.in_high_priority_pool = true;
+                self.high_priority_weight += record.weight();
                 self.high_priority_list.push_back(ptr);
 
                 self.may_overflow_high_priority_pool();
             }
-            LruContext::LowPriority => {
-                handle.in_high_priority_pool = false;
+            LruHint::LowPriority => {
+                record.state.in_high_priority_pool = false;
                 self.list.push_back(ptr);
             }
         }
 
-        handle.base_mut().set_in_eviction(true);
+        record.set_in_eviction(true);
     }
 
-    unsafe fn pop(&mut self) -> Option<NonNull<Self::Handle>> {
+    fn pop(&mut self) -> Option<NonNull<Record<Self>>> {
         let mut ptr = self.list.pop_front().or_else(|| self.high_priority_list.pop_front())?;
 
-        let handle = ptr.as_mut();
-        strict_assert!(!handle.link.is_linked());
+        let record = unsafe { ptr.as_mut() };
+        strict_assert!(!record.state.link.is_linked());
 
-        if handle.in_high_priority_pool {
-            self.high_priority_weight -= handle.base().weight();
-            handle.in_high_priority_pool = false;
+        if record.state.in_high_priority_pool {
+            self.high_priority_weight -= record.weight();
+            record.state.in_high_priority_pool = false;
         }
 
-        handle.base_mut().set_in_eviction(false);
+        record.set_in_eviction(false);
 
         Some(ptr)
     }
 
-    unsafe fn acquire(&mut self, _: NonNull<Self::Handle>) {}
+    fn remove(&mut self, mut ptr: NonNull<Record<Self>>) {
+        let record = unsafe { ptr.as_mut() };
+        strict_assert!(record.state.link.is_linked());
 
-    unsafe fn release(&mut self, mut ptr: NonNull<Self::Handle>) {
-        let handle = ptr.as_mut();
-
-        if handle.base().is_in_eviction() {
-            strict_assert!(handle.link.is_linked());
-            self.remove(ptr);
-            self.push(ptr);
+        if record.state.in_high_priority_pool {
+            self.high_priority_weight -= record.weight();
+            self.high_priority_list.remove(ptr);
+            record.state.in_high_priority_pool = false;
         } else {
-            strict_assert!(!handle.link.is_linked());
-            self.push(ptr);
+            self.list.remove(ptr);
         }
 
-        strict_assert!(handle.base().is_in_eviction());
+        strict_assert!(!record.state.link.is_linked());
+
+        record.set_in_eviction(false);
     }
 
-    unsafe fn remove(&mut self, mut ptr: NonNull<Self::Handle>) {
-        let handle = ptr.as_mut();
-        strict_assert!(handle.link.is_linked());
-
-        if handle.in_high_priority_pool {
-            self.high_priority_weight -= handle.base.weight();
-            self.high_priority_list.remove_raw(handle.link.raw());
-            handle.in_high_priority_pool = false;
-        } else {
-            self.list.remove_raw(handle.link.raw());
+    fn clear(&mut self) {
+        while let Some(mut ptr) = self.list.pop_front() {
+            unsafe { ptr.as_mut() }.set_in_eviction(false);
         }
-
-        strict_assert!(!handle.link.is_linked());
-
-        handle.base_mut().set_in_eviction(false);
-    }
-
-    unsafe fn clear(&mut self) -> Vec<NonNull<Self::Handle>> {
-        let mut res = Vec::with_capacity(self.len());
-
-        while !self.list.is_empty() {
-            let mut ptr = self.list.pop_front().strict_unwrap_unchecked();
-            ptr.as_mut().base_mut().set_in_eviction(false);
-            res.push(ptr);
-        }
-
-        while !self.high_priority_list.is_empty() {
-            let mut ptr = self.high_priority_list.pop_front().strict_unwrap_unchecked();
-            ptr.as_mut().base_mut().set_in_eviction(false);
-            ptr.as_mut().in_high_priority_pool = false;
-            self.high_priority_weight -= ptr.as_ref().base().weight();
-            res.push(ptr);
+        while let Some(mut ptr) = self.high_priority_list.pop_front() {
+            let record = unsafe { ptr.as_mut() };
+            record.set_in_eviction(false);
+            record.state.in_high_priority_pool = false;
+            self.high_priority_weight -= record.weight();
         }
 
         strict_assert_eq!(self.high_priority_weight, 0);
-
-        res
     }
 
     fn len(&self) -> usize {
         self.high_priority_list.len() + self.list.len()
     }
 
-    fn is_empty(&self) -> bool {
-        self.len() == 0
+    fn acquire_operator() -> super::Operator {
+        Operator::Immutable
+    }
+
+    fn acquire_immutable(&self, _ptr: NonNull<Record<Self>>) {}
+
+    fn acquire_mutable(&mut self, _ptr: NonNull<Record<Self>>) {
+        unreachable!()
+    }
+
+    fn release(&mut self, mut ptr: NonNull<Record<Self>>) {
+        let record = unsafe { ptr.as_mut() };
+
+        if record.is_in_eviction() {
+            strict_assert!(record.state.link.is_linked());
+            self.remove(ptr);
+            self.push(ptr);
+        } else {
+            strict_assert!(!record.state.link.is_linked());
+            self.push(ptr);
+        }
+
+        strict_assert!(record.is_in_eviction());
     }
 }
-
-unsafe impl<T> Send for Lru<T> where T: Send + Sync + 'static {}
-unsafe impl<T> Sync for Lru<T> where T: Send + Sync + 'static {}
 
 #[cfg(test)]
 pub mod tests {
@@ -294,44 +274,44 @@ pub mod tests {
     use itertools::Itertools;
 
     use super::*;
-    use crate::{eviction::test_utils::TestEviction, handle::HandleExt};
+    use crate::{eviction::test_utils::TestEviction, record::Data};
 
-    impl<T> TestEviction for Lru<T>
+    impl<K, V> TestEviction for Lru<K, V>
     where
-        T: Send + Sync + 'static + Clone,
+        K: Key + Clone,
+        V: Value + Clone,
     {
-        fn dump(&self) -> Vec<T> {
+        fn dump(&self) -> Vec<NonNull<Record<Self>>> {
             self.list
-                .iter()
-                .chain(self.high_priority_list.iter())
-                .map(|handle| handle.base().data_unwrap_unchecked().clone())
+                .iter_ptr()
+                .chain(self.high_priority_list.iter_ptr())
                 .collect_vec()
         }
     }
 
-    type TestLruHandle = LruHandle<u64>;
-    type TestLru = Lru<u64>;
+    type TestLru = Lru<u64, u64>;
 
-    unsafe fn new_test_lru_handle_ptr(data: u64, context: LruContext) -> NonNull<TestLruHandle> {
-        let mut handle = Box::<TestLruHandle>::default();
-        handle.init(0, data, 1, context);
+    unsafe fn new_test_lru_handle_ptr(data: u64, hint: LruHint) -> NonNull<Record<TestLru>> {
+        let handle = Box::new(Record::new(Data {
+            key: data,
+            value: data,
+            hint,
+            state: Default::default(),
+            hash: 0,
+            weight: 1,
+        }));
         NonNull::new_unchecked(Box::into_raw(handle))
     }
 
-    unsafe fn del_test_lru_handle_ptr(ptr: NonNull<TestLruHandle>) {
+    unsafe fn del_test_lru_handle_ptr(ptr: NonNull<Record<TestLru>>) {
         let _ = Box::from_raw(ptr.as_ptr());
     }
 
-    unsafe fn dump_test_lru(lru: &TestLru) -> (Vec<u64>, Vec<u64>) {
+    #[expect(clippy::type_complexity)]
+    unsafe fn dump_test_lru(lru: &TestLru) -> (Vec<NonNull<Record<TestLru>>>, Vec<NonNull<Record<TestLru>>>) {
         (
-            lru.list
-                .iter()
-                .map(|handle| *handle.base().data_unwrap_unchecked())
-                .collect_vec(),
-            lru.high_priority_list
-                .iter()
-                .map(|handle| *handle.base().data_unwrap_unchecked())
-                .collect_vec(),
+            lru.list.iter_ptr().collect_vec(),
+            lru.high_priority_list.iter_ptr().collect_vec(),
         )
     }
 
@@ -343,9 +323,9 @@ pub mod tests {
                     new_test_lru_handle_ptr(
                         i,
                         if i < 10 {
-                            LruContext::HighPriority
+                            LruHint::HighPriority
                         } else {
-                            LruContext::LowPriority
+                            LruHint::LowPriority
                         },
                     )
                 })
@@ -366,21 +346,27 @@ pub mod tests {
             assert_eq!(lru.len(), 4);
             assert_eq!(lru.high_priority_weight, 4);
             assert_eq!(lru.high_priority_list.len(), 4);
-            assert_eq!(dump_test_lru(&lru), (vec![], vec![0, 1, 2, 3]));
+            assert_eq!(dump_test_lru(&lru), (vec![], vec![ptrs[0], ptrs[1], ptrs[2], ptrs[3]]));
 
             // 0, [1, 2, 3, 4]
             lru.push(ptrs[4]);
             assert_eq!(lru.len(), 5);
             assert_eq!(lru.high_priority_weight, 4);
             assert_eq!(lru.high_priority_list.len(), 4);
-            assert_eq!(dump_test_lru(&lru), (vec![0], vec![1, 2, 3, 4]));
+            assert_eq!(
+                dump_test_lru(&lru),
+                (vec![ptrs[0]], vec![ptrs[1], ptrs[2], ptrs[3], ptrs[4]])
+            );
 
             // 0, 10, [1, 2, 3, 4]
             lru.push(ptrs[10]);
             assert_eq!(lru.len(), 6);
             assert_eq!(lru.high_priority_weight, 4);
             assert_eq!(lru.high_priority_list.len(), 4);
-            assert_eq!(dump_test_lru(&lru), (vec![0, 10], vec![1, 2, 3, 4]));
+            assert_eq!(
+                dump_test_lru(&lru),
+                (vec![ptrs[0], ptrs[10]], vec![ptrs[1], ptrs[2], ptrs[3], ptrs[4]])
+            );
 
             // 10, [1, 2, 3, 4]
             let p0 = lru.pop().unwrap();
@@ -388,21 +374,27 @@ pub mod tests {
             assert_eq!(lru.len(), 5);
             assert_eq!(lru.high_priority_weight, 4);
             assert_eq!(lru.high_priority_list.len(), 4);
-            assert_eq!(dump_test_lru(&lru), (vec![10], vec![1, 2, 3, 4]));
+            assert_eq!(
+                dump_test_lru(&lru),
+                (vec![ptrs[10]], vec![ptrs[1], ptrs[2], ptrs[3], ptrs[4]])
+            );
 
             // 10, [1, 3, 4]
             lru.remove(ptrs[2]);
             assert_eq!(lru.len(), 4);
             assert_eq!(lru.high_priority_weight, 3);
             assert_eq!(lru.high_priority_list.len(), 3);
-            assert_eq!(dump_test_lru(&lru), (vec![10], vec![1, 3, 4]));
+            assert_eq!(dump_test_lru(&lru), (vec![ptrs[10]], vec![ptrs[1], ptrs[3], ptrs[4]]));
 
             // 10, 11, [1, 3, 4]
             lru.push(ptrs[11]);
             assert_eq!(lru.len(), 5);
             assert_eq!(lru.high_priority_weight, 3);
             assert_eq!(lru.high_priority_list.len(), 3);
-            assert_eq!(dump_test_lru(&lru), (vec![10, 11], vec![1, 3, 4]));
+            assert_eq!(
+                dump_test_lru(&lru),
+                (vec![ptrs[10], ptrs[11]], vec![ptrs[1], ptrs[3], ptrs[4]])
+            );
 
             // 10, 11, 1, [3, 4, 5, 6]
             lru.push(ptrs[5]);
@@ -410,17 +402,29 @@ pub mod tests {
             assert_eq!(lru.len(), 7);
             assert_eq!(lru.high_priority_weight, 4);
             assert_eq!(lru.high_priority_list.len(), 4);
-            assert_eq!(dump_test_lru(&lru), (vec![10, 11, 1], vec![3, 4, 5, 6]));
+            assert_eq!(
+                dump_test_lru(&lru),
+                (
+                    vec![ptrs[10], ptrs[11], ptrs[1]],
+                    vec![ptrs[3], ptrs[4], ptrs[5], ptrs[6]]
+                )
+            );
 
             // 10, 11, 1, 3, [4, 5, 6, 0]
             lru.push(ptrs[0]);
             assert_eq!(lru.len(), 8);
             assert_eq!(lru.high_priority_weight, 4);
             assert_eq!(lru.high_priority_list.len(), 4);
-            assert_eq!(dump_test_lru(&lru), (vec![10, 11, 1, 3], vec![4, 5, 6, 0]));
+            assert_eq!(
+                dump_test_lru(&lru),
+                (
+                    vec![ptrs[10], ptrs[11], ptrs[1], ptrs[3]],
+                    vec![ptrs[4], ptrs[5], ptrs[6], ptrs[0]]
+                )
+            );
 
-            let ps = lru.clear();
-            assert_eq!(ps, [10, 11, 1, 3, 4, 5, 6, 0].map(|i| ptrs[i]));
+            lru.clear();
+            assert_eq!(dump_test_lru(&lru), (vec![], vec![]));
 
             for ptr in ptrs {
                 del_test_lru_handle_ptr(ptr);
