+++ foyer-memory/src/indexer/mod.rs
@@ -15,24 +15,21 @@
 use std::{hash::Hash, ptr::NonNull};
 
 use equivalent::Equivalent;
-use foyer_common::code::Key;
 
-use crate::handle::KeyedHandle;
+use crate::{eviction::Eviction, record::Record};
 
-pub trait Indexer: Send + Sync + 'static {
-    type Key: Key;
-    type Handle: KeyedHandle<Key = Self::Key>;
+pub trait Indexer: Send + Sync + 'static + Default {
+    type Eviction: Eviction;
 
-    fn new() -> Self;
-    unsafe fn insert(&mut self, ptr: NonNull<Self::Handle>) -> Option<NonNull<Self::Handle>>;
-    unsafe fn get<Q>(&self, hash: u64, key: &Q) -> Option<NonNull<Self::Handle>>
+    fn insert(&mut self, ptr: NonNull<Record<Self::Eviction>>) -> Option<NonNull<Record<Self::Eviction>>>;
+    fn get<Q>(&self, hash: u64, key: &Q) -> Option<NonNull<Record<Self::Eviction>>>
     where
-        Q: Hash + Equivalent<Self::Key> + ?Sized;
-    unsafe fn remove<Q>(&mut self, hash: u64, key: &Q) -> Option<NonNull<Self::Handle>>
+        Q: Hash + Equivalent<<Self::Eviction as Eviction>::Key> + ?Sized;
+    fn remove<Q>(&mut self, hash: u64, key: &Q) -> Option<NonNull<Record<Self::Eviction>>>
     where
-        Q: Hash + Equivalent<Self::Key> + ?Sized;
-    unsafe fn drain(&mut self) -> impl Iterator<Item = NonNull<Self::Handle>>;
+        Q: Hash + Equivalent<<Self::Eviction as Eviction>::Key> + ?Sized;
+    fn drain(&mut self) -> impl Iterator<Item = NonNull<Record<Self::Eviction>>>;
 }
 
 pub mod hash_table;
-pub mod sanity;
+pub mod sentry;
