+++ foyer-memory/src/record.rs
