+++ foyer-memory/src/slab.rs
